[ch24682]Move app-shell (#324)
* Move app-shell

* submodule ref

* submodule ref
