+++ index.html
@@ -165,6 +165,6 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
     <script src="node_modules/dayjs/plugin/isSameOrBefore.js"></script>
     <script src="node_modules/focus-visible/dist/focus-visible.min.js"></script>
     <!-- Built with love using PWA Starter Kit -->
-    <script type="module" src="src/components/app-shell/app-shell.js"></script>
+    <script type="module" src="src/app-shell.js"></script>
   </body>
 </html>
